requirements.txt
crewai>=0.28.0
langchain>=0.1.0
chonkie>=0.3.0
deepeval>=0.20.4
qdrant-client>=1.7.0
sentence-transformers>=2.2.2
openai>=1.3.0
python-dotenv>=1.0.0
pytest>=7.4.0
tqdm>=4.66.1
streamlit>=1.26.0
